python_4to/requirements.txt
wikipedia
gtts
opencv-python
pillow
pyaudio
chatGPT
requests
numpy
flask
mysql-connector-python
tkinter
openai
PatternFill
Alignment
messagebox
Workbook
ImageGrab
pytesseract
speech_recognition
serial


¿Y tu clave API?
Como ya mencionaste que la configuraste, asegúrate de que en asistente.py hayas hecho esto:

python
Copiar
Editar
openai.api_key = "sk-..."  # tu clave real de OpenAI
O bien, si usas un archivo .env, haz load_dotenv() como te enseñé antes.